R6 class for a cabinet

Constructs an R6 class of FileCabinet. Objects of class FileCabinet contain information that is used by new_cabinet_proj() to create project directories.

Public fields

name

cabinet name.

directory

the path to where future directories will be created, a string.

structure

the directory structure, a list.

Methods

Public methods

FileCabinet$new()

FileCabinet$print()

Method new()

Usage

FileCabinet$new(name, directory, structure)

Arguments

name

cabinet name.

directory

the path to where future directories will be created, a string.

structure

the directory structure, a list.

Details

Create a new `FileCabinet` object.

Returns

A cabinet object.

Examples

FileCabinet$new("test", "a/path", list(code = NULL, 'data/derived' = NULL, 'data/source' = NULL))

Method print()

Usage

FileCabinet$print()

Details

Print an object of class FileCabinet.

Examples

## ------------------------------------------------
## Method `FileCabinet$new`
## ------------------------------------------------

FileCabinet$new("test", "a/path", list(code = NULL, 'data/derived' = NULL, 'data/source' = NULL))

#> Cabinet name: test
#> Cabinet path: a/path
#> Cabinet structure: 
#> ├── code
#> └── data
#>     ├── derived
#>     └── source