R project settings

create_r_proj is a helper function for new_cabinet_proj. Calling it outside the scope of new_cabinet_proj will only print the specified settings the console.

create_r_proj(
  version = "1.0",
  restore_workspace = c("No", "Yes", "Default"),
  save_workspace = c("No", "Yes", "Default"),
  save_history = c("Default", "No", "Yes"),
  enable_code_indexing = c("Yes", "No", "Default"),
  spaces_for_tab = c("Yes", "No", "Default"),
  num_spaces_for_tab = 2,
  encoding = "UTF-8",
  rnw_weave = "Sweave",
  latex = "pdfLaTeX",
  auto_append_new_line = "Yes",
  strip_trailing_white_space = "Yes"
)

Arguments

version
R project version number, to be passed as character string.

restore_workspace
Load the .Rdata file (if any) found in the initial working directory into the R workspace. Options are "No" (default), "Yes", and "Default". If "Default", global behaviour settings are inherited.

save_workspace
Save .RData on exit. Options are "No" (default), "Yes", and "Default". If "Default", global behaviour settings are inherited.

save_history
Always save history (even when not saving .RData). Options are "Default" (default), "Yes", and "No". If "Default", global behaviour settings are inherited.

enable_code_indexing
Determines whether R source files within the project directory are indexed for code navigation. Options are "Yes" (default), "No", and "Default". If "Default", global behaviour settings are inherited.

spaces_for_tab
Determine whether the tab key inserts multiple spaces rather than a tab character (soft tabs). Options are "Yes" (default), "No", and "Default". If "Default", global behaviour settings are inherited.

num_spaces_for_tab
Specify the number of spaces per soft-tab, integer.

encoding
Specify the default text encoding for source files. Default is "UTF-8".

rnw_weave
Specify how to weave Rnw files. Default is "Sweave".

latex
Specify LaTeX processing. Default is "pdfLaTeX".

auto_append_new_line
Ensure that source files end with a new line. Default is "Yes".

strip_trailing_white_space
Strip trailing horizontal white space when saving. Default is "Yes".

Value

The settings used to write a .Rproj file.

See also

Examples

create_r_proj()

#> Version: 1.0
#> RestoreWorkspace: No
#> SaveWorkspace: No
#> AlwaysSaveHistory: Default
#> EnableCodeIndexing: Yes
#> UseSpacesForTab: Yes
#> NumSpacesForTab: 2
#> Encoding: UTF-8
#> RnwWeave: Sweave
#> LaTeX: pdfLaTeX
#> AutoAppendNewline: Yes
#> StripTrailingWhitespace: Yes